pyproject.toml
[project]
name = "arxiv2bib"
version = "0.1.0"
description = "A CLI utility to fetch BibTeX citations from arXiv IDs."
authors = [{ name = "Ihor Omelchenko", email = "ihor.v.omelchenko@gmail.com" }]
dependencies = [
    "click",
    "arxiv",
    "requests",
    "bibtexparser",
    "rich",
    "crossrefapi",
    "click",
]

[project.scripts]
arxiv2bib = "arxiv2bib.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["notebooks"]
